moveed selection dialog in another file named CurrencySelectionDialog.kt
